Refactor FocusGameViewAttribute
